Ajuste del enpoint POST asignar tareas a ETIQUETAS. Cambiar la direcion correcta de la url en los test_tasks_filters.py y test_tags.py.
