+++ app/api/v1/endpoints/tasks.py
@@ -11,7 +11,6 @@ from sqlalchemy.exc import IntegrityError
 from sqlalchemy import desc, asc
 from app.db.session import get_session
 from app.models import CategoriaTarea, TaskTag
-from app.schemas.tag import TagAssignRequest
 from app.services.auth import get_current_user
 from app.models.task import Task, TaskHistory, EstadoTarea
 from app.schemas.task import TaskCreate, TaskRead, TaskUpdate, TaskAssignmentCreate, TaskAssignmentRead
@@ -227,25 +226,3 @@ async def get_assigned_tasks(
 ):
     return await TaskAssignmentService.get_assigned_tasks(session=session, user_id=user_id)
 
-@router.post("/{task_id}/tags", status_code=200, summary="Asignar etiquetas", description="Asigna etiquetas a una tarea específica.")
-async def assign_tags_to_task(
-        task_id: UUID,
-        payload: TagAssignRequest,
-        session: AsyncSession = Depends(get_session),
-        current_user: Usuario = Depends(get_current_user)
-):
-    # Verifica que la tarea exista y pertenezca al usuario
-    task = await session.get(Task, task_id)
-    if not task or task.user_id != current_user.id:
-        raise HTTPException(status_code=404, detail="Tarea no encontrada.")
-
-    # Crear relaciones TaskTag (evitar duplicados)
-    for tag_id in payload.tag_ids:
-        result = await session.exec(
-            select(TaskTag).where(TaskTag.task_id == task_id, TaskTag.tag_id == tag_id)
-        )
-        if not result.first():
-            session.add(TaskTag(task_id=task_id, tag_id=tag_id))
-
-    await session.commit()
-    return {"message": "Etiquetas asignadas correctamente"}
